ITS-13: Project structure
